+++ CHANGELOG.md
@@ -2,7 +2,6 @@
 
 ## 4.3.0 - Unreleased
 ### Added
-- Added a cached includes column to the Blitz cache utility for the File Cache Storage driver. 
 - Added the `ssiEnabled` config setting that enables Blitz to include templates using Server Side Includes (SSI), which must be enabled on the web server.
 - Added the `esiEnabled` config setting that enables Blitz to include templates using Edge Side Includes (ESI), which must be enabled on the web server or reverse proxy (CDN).
 - Added the `craft.blitz.include()` template variable, that includes a template using SSI if enabled, otherwise via an AJAX request. The `include()` method returns a cached result (if one exists).
@@ -10,6 +9,8 @@
 - Added the `craft.blitz.fetch()` template variable, that fetches a URI via an AJAX request. Whether the URI response is cached or not is determined by the URI patterns in the plugin settings.
 - Added the `blitz/templates/include` and `blitz/templates/dynamic-include` controller actions.
 - Added the `maxUriLength` config setting.
+- Added a `rewrite.php` file that can be used in situations where a server rewrite is not possible.
+- Added a cached includes column to the Blitz cache utility for the File Cache Storage driver. 
 
 ### Changed
 - Improved the detection of when elements should be refreshed based on changes.